bot-telegram/backend-py/requirements.txt
fastapi==0.120.4
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.2.2
